feat: add ServiceSelection component
- Create dedicated ServiceSelection component
- Handle service type and options selection
- Improve component organization
- Add proper prop types and handlers
- Maintain responsive design
- Improve code reusability
