+++ src/components/pages/projet/form/components/DetailsInput.jsx
@@ -1,0 +1,4 @@
+export default function DetailsInput() {
+  return (
+    
+  )
